* Linux/386/include/lib9.h: update Rune type! so can compile plan9 programs and not get typechecking error when use L"whatever"

+++ Linux/386/include/lib9.h
@@ -39,7 +39,9 @@ typedef struct Proc Proc;
 
 typedef unsigned char	uchar;
 typedef signed char	schar;
-typedef unsigned short	Rune;
+//PAD: typedef ushort		Rune; backported from latest plan9
+// to not get some "illegal rune" error when compiling libc/
+typedef unsigned int	Rune;
 typedef long long int	vlong;
 typedef unsigned long long int	uvlong;
 typedef unsigned int u32int;
